requirements.txt
Flask==3.1.2
web3==7.14.0
python-dotenv==1.2.1
py-solc-x==2.0.5
